# src/main/java/ru/jafix/springproject/dto/common/StatusDto.java
package ru.jafix.springproject.dto.common;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class StatusDto {
    private Status status;
    private String message;
}